Front-end -> Integração: Integrando com Backend - crud completo, atualizando tanto database quanto view...

+++ front-end/src/js/controller/ExpenseController.ts
@@ -19,39 +19,54 @@ class ExpenseController {
         this._expenseListView.update(this._expenseListModel)
     }
 
-    public async get(){
+    public async get(): Promise<void> {
 
         await ExpenseServices.list()
-            .then(list =>  list.forEach((expense: ExpenseModel) => this._expenseListModel.add(expense)))
+            .then(list => list.forEach((expense: ExpenseModel) => this._expenseListModel.add(expense)))
             .catch(error => console.log(error))
 
         this.updateView()
     }
 
-    public create(): void {
-        const expense = this.createExpense()
-    
-        this._expenseListModel.add(expense)
+    public async create(): Promise<void> {
+
+        await ExpenseServices.add(this.createExpense())
+            .then((expense: ExpenseModel) => this._expenseListModel.add(expense))
+            .catch(error => console.log(error))
+
         this.updateView()
         this.resetForm()
         
     }
 
-    public update(): void {
-        const expense = this.setUpdatedExpense()
-        const updated = this._expenseListModel.update(expense)
+    public async update(): Promise<void> {
 
-        if(!updated){
-            throw new Error('Não foi possivel atualizar dados de despesa!')
-            return
-        }
+        await ExpenseServices.update(this.setUpdatedExpense())
+            .then((expense: ExpenseModel) => {
+                const updated = this._expenseListModel.update(expense)
+                if(!updated){
+                    throw new Error('Não foi possivel atualizar dados de despesa!')
+                    return
+                }
+            })
+            .catch(error => console.log(error))
 
         this.updateView()
         this.resetForm()
     }
 
-    public delete(id: number): void {
-        this._expenseListModel.delete(id)
+    public async delete(id: number): Promise<void> {
+
+        await ExpenseServices.delete(id)
+            .then((deleted: boolean) => {
+                if(!deleted){
+                    throw new Error('Não foi possivel deletar dados de despesa!')
+                    return
+                }
+                this._expenseListModel.delete(id)
+            })
+            .catch(error => console.log(error))
+
         this.updateView()
         this.resetForm()
     }
@@ -87,8 +102,6 @@ class ExpenseController {
 
     private updateView(): void {
         this._expenseListView.update(this._expenseListModel)
-
-        console.log( this._expenseListModel.list() )
     }
 
     private setCurrencyFormat(value: string): number {
